Fix NerdTreeFind issue
With a new created file that is not yet on the NERDTree node.
When invoking NerdTreeFind on that file it will raise an error because
of that file not being located in node nerdtree. This issue was raised
in NERDTree repo but still not fixed. This is a temporary fix since it
just mitigates the error by refreshing the node.

https://github.com/scrooloose/nerdtree/issues/779